studentoverflow/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate ==4.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1